(MOT-4107) refactor(harness): centralize integration runtime ownership
