+++ harness/evals/integration/src/stack.rs
@@ -1,457 +1,27 @@
-//! Per-scenario stack supervision (spec § Supervisor contract): run-dir
-//! allocation, engine/seed YAML rendering, child spawning in declared order
-//! with an environment allowlist and per-process log capture, early-exit
-//! detection, and SIGTERM→SIGKILL teardown.
+//! Per-scenario stack planning and supervision.
 //!
-//! Engine facts verified against the pinned source (iii-hq/iii @ 085e0fde):
-//! - `iii-worker-manager` accepts `config: { port, host }` in the engine
-//!   YAML; mandatory builtins merge in even when a config file is used
-//!   (`engine/src/workers/config.rs`), and a listed entry's config wins.
-//! - `configuration` (engine builtin) accepts
-//!   `adapter: { name: fs, config: { directory } }`.
-//! - `enabled_by_default` builtins (`iii-state`, `iii-stream`, `iii-cron`)
-//!   are NOT auto-added when a config file is used — they must be listed.
+//! Engine facts verified against the pinned source:
+//! - `iii-worker-manager` accepts `config: { port, host }`;
+//! - `configuration` accepts an fs adapter directory;
+//! - enabled-by-default builtins must be listed when a config file is used.
 
-use std::collections::BTreeMap;
-use std::path::{Path, PathBuf};
-use std::time::Duration;
+mod bins;
+mod config;
+mod layout;
+mod manifest;
+mod supervisor;
 
-use anyhow::Context;
-use serde_json::{json, Value};
-
-use crate::process::{ProcessSpec, ProcessSupervisor, DEFAULT_TEARDOWN_BUDGET};
+#[cfg(test)]
+mod tests;
 
 pub use crate::process::EarlyExit;
+pub use bins::StackBins;
+pub use config::{
+    expected_config_entries, expected_harness_config_entry, render_engine_yaml, render_seed,
+    WORKER_START_ORDER,
+};
+pub use layout::RunLayout;
+pub use supervisor::{free_loopback_port, Stack};
 
-/// Start order matters: the harness retries `queue::define` only briefly at
-/// boot (`harness/src/queue.rs`), so queue precedes harness. The engine is
-/// always first. The harness itself is spawned LAST, by an explicit
-/// [`Stack::spawn_harness`] call after Arm: its native-exposure registry
-/// snapshot is seeded at boot and refreshed reactively
-/// (`harness/src/discovery.rs`), so the run-scoped recorder target must be
-/// registered before the harness starts or the first turn races the refresh.
-pub const WORKER_START_ORDER: [&str; 4] = [
-    "queue",
-    "iii-directory",
-    "session-manager",
-    "context-manager",
-];
-
-/// Environment allowlist for every child (spec step 3): no provider keys or
-/// developer secrets leak into the subject stack.
-const ENV_ALLOWLIST: [&str; 4] = ["PATH", "HOME", "LANG", "RUST_LOG"];
-
-#[derive(Debug, Clone)]
-pub struct StackBins {
-    pub engine: PathBuf,
-    pub harness: PathBuf,
-    /// queue, iii-directory, session-manager, context-manager.
-    pub workers: BTreeMap<String, PathBuf>,
-}
-
-impl StackBins {
-    /// Every binary the scenario spawns, in declared start order.
-    pub fn resolve(&self, name: &str) -> Option<&Path> {
-        match name {
-            "engine" => Some(&self.engine),
-            "harness" => Some(&self.harness),
-            other => self.workers.get(other).map(PathBuf::as_path),
-        }
-    }
-
-    pub fn missing_workers(&self) -> Vec<&'static str> {
-        WORKER_START_ORDER
-            .iter()
-            .filter(|name| !self.workers.contains_key(**name))
-            .copied()
-            .collect()
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct RunPaths {
-    pub root: PathBuf,
-    pub engine_dir: PathBuf,
-    pub seeds_dir: PathBuf,
-    pub logs_dir: PathBuf,
-}
-
-impl RunPaths {
-    pub fn allocate(artifacts_dir: &Path, run_id: &str) -> anyhow::Result<Self> {
-        let root = artifacts_dir.join(run_id);
-        let paths = Self {
-            engine_dir: root.join("engine"),
-            seeds_dir: root.join("seeds"),
-            logs_dir: root.join("logs"),
-            root,
-        };
-        for dir in [
-            &paths.root,
-            &paths.engine_dir,
-            &paths.seeds_dir,
-            &paths.logs_dir,
-        ] {
-            std::fs::create_dir_all(dir).with_context(|| format!("creating {}", dir.display()))?;
-        }
-        Ok(paths)
-    }
-
-    pub fn scenario_dir(&self, scenario_id: &str) -> anyhow::Result<PathBuf> {
-        crate::types::scenario::validate_scenario_id(scenario_id)?;
-        let dir = self.root.join("scenarios").join(scenario_id);
-        std::fs::create_dir_all(&dir).with_context(|| format!("creating {}", dir.display()))?;
-        Ok(dir)
-    }
-}
-
-/// A free loopback port, found by bind-and-release. Racy by nature; the boot
-/// path retries once with a fresh port on a bind failure (spec step 2).
-pub fn free_loopback_port() -> anyhow::Result<u16> {
-    let listener = std::net::TcpListener::bind(("127.0.0.1", 0))?;
-    Ok(listener.local_addr()?.port())
-}
-
-pub fn render_engine_yaml(paths: &RunPaths, port: u16) -> String {
-    // `${` would trip the engine's env expansion; none of these values can
-    // contain it (ports and run-dir paths).
-    format!(
-        r#"# Generated by harness-integration; one engine per scenario run.
-workers:
-  - name: iii-worker-manager
-    config:
-      host: 127.0.0.1
-      port: {port}
-  - name: configuration
-    config:
-      adapter:
-        name: fs
-        config:
-          directory: {config_dir}
-  - name: iii-state
-  - name: iii-stream
-  - name: iii-cron
-"#,
-        config_dir = paths.engine_dir.join("configuration").display(),
-    )
-}
-
-/// Seed YAML per configurable worker (spec step 4). Shapes verified against
-/// each worker's config module.
-pub fn render_seed(worker: &str, paths: &RunPaths) -> Option<Value> {
-    match worker {
-        "session-manager" => Some(json!({
-            "adapter": {
-                "name": "fs",
-                "config": { "data_dir": paths.root.join("session-data").to_string_lossy() }
-            }
-        })),
-        "context-manager" => Some(json!({
-            "lease_dir": paths.root.join("leases").to_string_lossy()
-        })),
-        "queue" => Some(json!({
-            "adapter": {
-                "name": "builtin",
-                "config": {
-                    "store_method": "file_based",
-                    "file_path": paths.root.join("queue").to_string_lossy(),
-                    "save_interval_ms": 1000
-                }
-            }
-        })),
-        "iii-directory" => Some(json!({
-            "skills_folder": paths.root.join("skills").to_string_lossy()
-        })),
-        // Without this the harness scopes each new session to its boot cwd
-        // (the run root) and appends a run-path-dependent "Your working
-        // directory is …" line to the system prompt — which would make the
-        // expected-prompt sha256 depend on the artifacts path.
-        "harness" => Some(json!({
-            "default_filesystem_root": "off"
-        })),
-        _ => None,
-    }
-}
-
-/// Expected `configuration::get` entries after the pre-harness boot: worker
-/// config id → seeded value (readiness asserts each seeded key is
-/// authoritative in the stored resolved config).
-pub fn expected_config_entries(paths: &RunPaths) -> Vec<(String, Value)> {
-    WORKER_START_ORDER
-        .iter()
-        .filter_map(|worker| render_seed(worker, paths).map(|seed| (worker.to_string(), seed)))
-        .collect()
-}
-
-/// The harness's own seeded entry, probed after [`Stack::spawn_harness`].
-pub fn expected_harness_config_entry(paths: &RunPaths) -> Vec<(String, Value)> {
-    render_seed("harness", paths)
-        .map(|seed| vec![("harness".to_string(), seed)])
-        .unwrap_or_default()
-}
-
-pub struct Stack {
-    pub ws_url: String,
-    pub paths: RunPaths,
-    processes: ProcessSupervisor,
-    /// Engine spawn recipe, kept for fault-injection respawns:
-    /// (binary, args, cwd). `None` only in test stacks.
-    engine_recipe: Option<(PathBuf, Vec<String>, PathBuf)>,
-    /// How many times the engine was intentionally killed and respawned
-    /// (suffixes the respawned process's log files).
-    engine_restarts: u32,
-}
-
-impl Stack {
-    /// Spawn the engine and every worker in declared order. This does not
-    /// wait for readiness — `readiness::probe` owns that. A bind race
-    /// (engine exits early because the chosen port got taken) is retried
-    /// once with a completely new port set.
-    pub async fn boot(bins: &StackBins, paths: RunPaths) -> anyhow::Result<Stack> {
-        match Self::boot_once(bins, &paths).await {
-            Ok(stack) => Ok(stack),
-            Err(BootError::BindRace) => {
-                tracing::warn!("engine bind race; retrying with a fresh port set");
-                Self::boot_once(bins, &paths)
-                    .await
-                    .map_err(|e| e.into_anyhow())
-            }
-            Err(e) => Err(e.into_anyhow()),
-        }
-    }
-
-    async fn boot_once(bins: &StackBins, paths: &RunPaths) -> Result<Stack, BootError> {
-        let missing = bins.missing_workers();
-        if !missing.is_empty() {
-            return Err(BootError::Other(anyhow::anyhow!(
-                "missing --worker-bin for: {}",
-                missing.join(", ")
-            )));
-        }
-
-        let port = free_loopback_port().map_err(BootError::Other)?;
-        let ws_url = format!("ws://127.0.0.1:{port}");
-
-        let engine_yaml_path = paths.engine_dir.join("engine.yaml");
-        std::fs::write(&engine_yaml_path, render_engine_yaml(paths, port))
-            .map_err(|e| BootError::Other(e.into()))?;
-
-        let info = stack_info(bins, paths, port);
-        std::fs::write(
-            paths.root.join("stack.json"),
-            crate::canonical::canonical_json_pretty(&info),
-        )
-        .map_err(|e| BootError::Other(e.into()))?;
-
-        let engine_args = vec![
-            "--config".to_string(),
-            engine_yaml_path.to_string_lossy().into_owned(),
-            "--no-update-check".to_string(),
-        ];
-        let mut stack = Stack {
-            ws_url,
-            paths: paths.clone(),
-            processes: ProcessSupervisor::default(),
-            engine_recipe: Some((
-                bins.engine.clone(),
-                engine_args.clone(),
-                paths.engine_dir.clone(),
-            )),
-            engine_restarts: 0,
-        };
-
-        stack
-            .spawn_child("engine", &bins.engine, &engine_args, &paths.engine_dir)
-            .map_err(BootError::Other)?;
-
-        // Give the engine a moment to bind, then classify an immediate death
-        // as a bind race exactly once (any other exit reason surfaces via the
-        // readiness probe's early-exit check with full logs).
-        tokio::time::sleep(std::time::Duration::from_millis(600)).await;
-        if let Some(exit) = stack.early_exit() {
-            stack.teardown().await;
-            if exit.name == "engine" {
-                return Err(BootError::BindRace);
-            }
-            return Err(BootError::Other(anyhow::anyhow!(
-                "{} exited during boot: {}",
-                exit.name,
-                exit.status
-            )));
-        }
-
-        for worker in WORKER_START_ORDER {
-            let bin = bins
-                .resolve(worker)
-                .expect("presence checked above")
-                .to_path_buf();
-            stack.spawn_worker(worker, &bin).map_err(BootError::Other)?;
-        }
-
-        Ok(stack)
-    }
-
-    /// Spawn the harness under test. Called after Arm so the run-scoped
-    /// target is already registered when the harness seeds its registry
-    /// snapshot (see [`WORKER_START_ORDER`]).
-    pub fn spawn_harness(&mut self, bins: &StackBins) -> anyhow::Result<()> {
-        let bin = bins.harness.clone();
-        self.spawn_worker("harness", &bin)
-    }
-
-    /// Fault injection (`fault.kind: engine_sigkill`): SIGKILL the engine —
-    /// no shutdown grace, mirroring a hard crash — and reap it so
-    /// `early_exit` does not classify the intentional kill as a
-    /// `process_crash`. Workers keep running; their SDK connections retry
-    /// and replay registrations once the engine returns.
-    pub async fn kill_engine(&mut self) -> anyhow::Result<()> {
-        let mut engine = self
-            .processes
-            .remove("engine")
-            .ok_or_else(|| anyhow::anyhow!("no live engine child to kill"))?;
-        engine.kill_now().await?;
-        tracing::info!("engine SIGKILLed (fault injection)");
-        Ok(())
-    }
-
-    /// Respawn the engine from the recipe recorded at boot: same YAML, same
-    /// working directory, same port — the restart boundary the scenario is
-    /// probing. Logs land in `engine.restart<N>.{out,err}` so the crashed
-    /// process's output survives.
-    pub fn respawn_engine(&mut self) -> anyhow::Result<()> {
-        let (bin, args, cwd) = self
-            .engine_recipe
-            .clone()
-            .ok_or_else(|| anyhow::anyhow!("stack has no engine recipe (test stack?)"))?;
-        self.engine_restarts += 1;
-        let log_name = format!("engine.restart{}", self.engine_restarts);
-        self.spawn_child_logged("engine", &log_name, &bin, &args, &cwd)
-    }
-
-    fn spawn_worker(&mut self, worker: &str, bin: &Path) -> anyhow::Result<()> {
-        let paths = self.paths.clone();
-        let mut args = vec!["--url".to_string(), self.ws_url.clone()];
-        if let Some(seed) = render_seed(worker, &paths) {
-            let seed_path = paths.seeds_dir.join(format!("{worker}.yaml"));
-            let yaml = serde_yaml::to_string(&seed)?;
-            std::fs::write(&seed_path, yaml)?;
-            args.push("--config".to_string());
-            args.push(seed_path.to_string_lossy().into_owned());
-        }
-        self.spawn_child(worker, bin, &args, &paths.root)
-    }
-
-    /// Test support: an empty stack over pre-allocated run paths, letting
-    /// supervisor tests drive `spawn_child`/`early_exit`/`teardown` with stub
-    /// child processes instead of a real engine.
-    #[doc(hidden)]
-    pub fn for_tests(paths: RunPaths) -> Stack {
-        Stack {
-            ws_url: "ws://127.0.0.1:0".to_string(),
-            paths,
-            processes: ProcessSupervisor::new(DEFAULT_TEARDOWN_BUDGET),
-            engine_recipe: None,
-            engine_restarts: 0,
-        }
-    }
-
-    /// Test support with a short teardown budget.
-    #[doc(hidden)]
-    pub fn for_tests_with_teardown_budget(paths: RunPaths, teardown_budget: Duration) -> Stack {
-        let mut stack = Self::for_tests(paths);
-        stack.set_teardown_budget(teardown_budget);
-        stack
-    }
-
-    /// Configure the total time available for SIGTERM grace, SIGKILL
-    /// escalation and reaping.
-    pub fn set_teardown_budget(&mut self, teardown_budget: Duration) {
-        self.processes.set_teardown_budget(teardown_budget);
-    }
-
-    pub fn teardown_budget(&self) -> Duration {
-        self.processes.teardown_budget()
-    }
-
-    #[doc(hidden)]
-    pub fn spawn_child(
-        &mut self,
-        name: &str,
-        bin: &Path,
-        args: &[String],
-        cwd: &Path,
-    ) -> anyhow::Result<()> {
-        self.spawn_child_logged(name, name, bin, args, cwd)
-    }
-
-    fn spawn_child_logged(
-        &mut self,
-        name: &str,
-        log_name: &str,
-        bin: &Path,
-        args: &[String],
-        cwd: &Path,
-    ) -> anyhow::Result<()> {
-        let stdout_log = self.paths.logs_dir.join(format!("{log_name}.out"));
-        let stderr_log = self.paths.logs_dir.join(format!("{log_name}.err"));
-        let mut spec =
-            ProcessSpec::new(name, bin, cwd, stdout_log, stderr_log).args(args.iter().cloned());
-        for key in ENV_ALLOWLIST {
-            if let Ok(value) = std::env::var(key) {
-                spec = spec.env(key, value);
-            }
-        }
-        self.processes.spawn(spec).map(|_| ())
-    }
-
-    /// First child that has already exited, if any (spec step 7: early
-    /// process exit outranks any ordinary timeout).
-    pub fn early_exit(&mut self) -> Option<EarlyExit> {
-        self.processes.early_exit()
-    }
-
-    /// SIGTERM everything (children first, engine last), then SIGKILL
-    /// remaining process groups within the configured teardown budget.
-    pub async fn teardown(&mut self) {
-        self.processes.teardown().await;
-    }
-}
-
-enum BootError {
-    BindRace,
-    Other(anyhow::Error),
-}
-
-impl BootError {
-    fn into_anyhow(self) -> anyhow::Error {
-        match self {
-            BootError::BindRace => anyhow::anyhow!("engine failed to bind twice in a row"),
-            BootError::Other(e) => e,
-        }
-    }
-}
-
-fn stack_info(bins: &StackBins, paths: &RunPaths, port: u16) -> Value {
-    let mut binaries = serde_json::Map::new();
-    let mut record = |name: &str, path: &Path| {
-        let absolute = path.canonicalize().unwrap_or_else(|_| path.to_path_buf());
-        let digest = std::fs::read(&absolute)
-            .map(|bytes| crate::canonical::sha256_of_bytes(&bytes))
-            .unwrap_or_else(|e| format!("unreadable: {e}"));
-        binaries.insert(
-            name.to_string(),
-            json!({ "path": absolute.to_string_lossy(), "sha256": digest }),
-        );
-    };
-    record("engine", &bins.engine);
-    record("harness", &bins.harness);
-    for (name, path) in &bins.workers {
-        record(name, path);
-    }
-    json!({
-        "port": port,
-        "run_root": paths.root.to_string_lossy(),
-        "binaries": Value::Object(binaries),
-        "env_allowlist": ENV_ALLOWLIST,
-        "start_order": WORKER_START_ORDER,
-    })
-}
+/// Compatibility name retained for callers of the original stack API.
+pub type RunPaths = RunLayout;
